Move Footer component to SimpleLayout
